Event table example with {gt} (#84)
* Disposition table with {gt}

Layout differs from flextable and tables example. It highlights the fact that Death scenarios are only subcategories for subjects who discontinued due to death. To be discussed.

* event table with {gt}

Added event table section for {gt}

* Indentation for ex tbl

* Update of gt demo table

Changed the following:
- added .groups = "drop" to dplyr::summarize to avoid messages
- updated syntax to native pipe
- moved summary row to top of the group

* Update of gt conmed table

added `.groups = "drop"` to calls to dplyr::summarize to avoid messages.

* Update of gt KM table

Changed syntax to native pipe and changed "." syntax
Added `.groups = "drop"` to calls to dplyr::summarize

* rendered book

* Removing remaining calls to `html()`

Replaced calls to `gt::html()` with `gt::md()`

* Including github version of gt

This might be deleted later and reverted to the original CRAN installation, but latest updates to the package are required to render the book.

* Temp update to demo table with gt

Commented out `side = "top"` argument, as not supported by current version. Tbc later.